remove readme and version in setup
