+++ README.md
@@ -26,11 +26,11 @@ SPDX-License-Identifier: MPL-2.0
 For detailed documentation, see [Read the Docs](https://power-grid-model-io.readthedocs.io/en/stable/index.html).
 
 # Examples
-* [PGM JSON Example](docs/examples/pgm_json_example.ipynb)
+* [PGM JSON Example](https://github.com/PowerGridModel/power-grid-model-io/tree/main/docs/examples)
 
 
 # License
-This project is licensed under the Mozilla Public License, version 2.0 - see [LICENSE](LICENSE) for details.
+This project is licensed under the Mozilla Public License, version 2.0 - see [LICENSE](https://github.com/PowerGridModel/power-grid-model-io/blob/main/LICENSE) for details.
 
 # Contributing
 Please read [CODE_OF_CONDUCT](https://github.com/PowerGridModel/.github/blob/main/CODE_OF_CONDUCT.md) and [CONTRIBUTING](https://github.com/PowerGridModel/.github/blob/main/CONTRIBUTING.md) for details on the process
