+++ pyproject.toml
@@ -5,8 +5,7 @@
 [build-system]
 requires = [
     "setuptools", 
-    "wheel", 
-    "requests",
+    "wheel"
 ]
 build-backend = "setuptools.build_meta"
